Add files via upload
The first part of the project successfully simulates a discrete time system with process noise.
The second part estimates system parameters using given inputs and outputs through the Recursive Least Squares estimation algorithm. It then compares the outputs of the estimated system (using the same inputs) to that of the originally given outputs.